Filter full functionality
Got the filter functionality working completely. prior to this the entire column of times was being accessed. Now the individual cell of times is displayed. Only major tasks remaining on this page are UI and formatting related, number of filters could be expanded on if there's time. Also made some UI changes to reviewFacility.